requirements.txt
# code management
black==23.10.1
coloredlogs==15.0.1
mypy==1.7.1
pydantic==2.5.2
pylint==3.0.0
pylint-exit==1.2.0
python-dotenv==1.0.0
memory_profiler==0.61.0

# pipelines
joblib==1.3.2
tqdm==4.66.3
wandb==0.15.12
transformers==4.46.2 # for args
pillow==10.2.0 # for images
datasets==3.1.0 # for downstream datasets
bitsandbytes==0.45.0

# data processing
numpy==1.26.0
pandas==2.2.0
scipy==1.11.4
scikit-learn==1.2.2
seaborn==0.13.0
statsmodels==0.14.1
matplotlib==3.8.2

# ai models
minicons==0.2.50
accelerate==0.31.0
huggingface-hub==0.24.0
tokenizers==0.20.3